+++ src/components/Budget/BudgetForm/BudgetForm.js
@@ -22,8 +22,8 @@ import {
   FormFooter,
 } from '@folio/stripes-acq-components';
 
+import FinancialSummary from '../../../common/FinancialSummary';
 import { SECTIONS_BUDGET } from '../constants';
-import BudgetSummary from '../BudgetView/BudgetSummary';
 import BudgetInformationFields from './BudgetInformationFields';
 import BudgetExpenseClassesFields from './BudgetExpenseClassesFields';
 
@@ -38,7 +38,6 @@ const BudgetForm = ({
   values,
 }) => {
   const { periodStart, periodEnd, currency } = fiscalYear;
-  const { awaitingPayment, encumbered, expended } = initialValues;
 
   const paneFooter = (
     <FormFooter
@@ -98,8 +97,8 @@ const BudgetForm = ({
                     id={SECTIONS_BUDGET.SUMMARY}
                   >
                     <ViewMetaData metadata={initialValues.metadata} />
-                    <BudgetSummary
-                      budget={initialValues}
+                    <FinancialSummary
+                      data={initialValues}
                       fiscalYearCurrency={currency}
                     />
                   </Accordion>
@@ -108,12 +107,8 @@ const BudgetForm = ({
                     id={SECTIONS_BUDGET.INFORMATION}
                   >
                     <BudgetInformationFields
-                      awaitingPayment={awaitingPayment}
-                      encumbered={encumbered}
-                      expended={expended}
                       fiscalEnd={periodEnd}
                       fiscalStart={periodStart}
-                      fiscalYearCurrency={currency}
                     />
                   </Accordion>
                   <Accordion