+++ src/components/Budget/BudgetView/BudgetInformation.js
@@ -10,7 +10,6 @@ import {
   Row,
 } from '@folio/stripes/components';
 import {
-  AmountWithCurrencyField,
   FolioFormattedDate,
 } from '@folio/stripes-acq-components';
 
@@ -18,22 +17,14 @@ import {
   TRANSACTIONS_ROUTE,
 } from '../../../common/const';
 
-const isOverAllowable = value => value > 100;
-
 const BudgetInformation = ({
   allowableEncumbrance,
   allowableExpenditure,
-  awaitingPayment,
   budgetStatus,
-  encumbered,
-  expenditures,
   fiscalEnd,
   fiscalStart,
-  fiscalYearCurrency,
   name,
   id,
-  overEncumbrance,
-  overExpended,
 }) => (
   <Row>
     <Col xs={3}>
@@ -54,6 +45,13 @@ const BudgetInformation = ({
       <KeyValue
         label={<FormattedMessage id="ui-finance.budget.status" />}
         value={budgetStatus}
+      />
+    </Col>
+
+    <Col xs={3}>
+      <KeyValue
+        label={<FormattedMessage id="ui-finance.budget.actionsRequired" />}
+        value={<NoValue />}
       />
     </Col>
 
@@ -87,76 +85,6 @@ const BudgetInformation = ({
 
     <Col xs={3}>
       <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.encumbered" />}
-      >
-        <AmountWithCurrencyField
-          amount={encumbered}
-          currency={fiscalYearCurrency}
-        />
-      </KeyValue>
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.awaitingPayment" />}
-      >
-        <AmountWithCurrencyField
-          amount={awaitingPayment}
-          currency={fiscalYearCurrency}
-        />
-      </KeyValue>
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.expended" />}
-      >
-        <AmountWithCurrencyField
-          amount={expenditures}
-          currency={fiscalYearCurrency}
-        />
-      </KeyValue>
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.overEncumbrance" />}
-      >
-        {
-          isOverAllowable(allowableEncumbrance) ? (
-            <AmountWithCurrencyField
-              amount={overEncumbrance}
-              currency={fiscalYearCurrency}
-            />
-          ) : <NoValue />
-        }
-      </KeyValue>
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.overExpended" />}
-      >
-        {
-          isOverAllowable(allowableExpenditure) ? (
-            <AmountWithCurrencyField
-              amount={overExpended}
-              currency={fiscalYearCurrency}
-            />
-          ) : <NoValue />
-        }
-      </KeyValue>
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
-        label={<FormattedMessage id="ui-finance.budget.actionsRequired" />}
-        value={<NoValue />}
-      />
-    </Col>
-
-    <Col xs={3}>
-      <KeyValue
         label={<FormattedMessage id="ui-finance.budget.transactions" />}
       >
         <Link to={`${TRANSACTIONS_ROUTE}/budget/${id}`}>
@@ -170,17 +98,11 @@ const BudgetInformation = ({
 BudgetInformation.propTypes = {
   allowableEncumbrance: PropTypes.number,
   allowableExpenditure: PropTypes.number,
-  awaitingPayment: PropTypes.number,
   budgetStatus: PropTypes.string,
-  encumbered: PropTypes.number,
-  expenditures: PropTypes.number,
   fiscalEnd: PropTypes.string,
   fiscalStart: PropTypes.string,
-  fiscalYearCurrency: PropTypes.string,
   name: PropTypes.string,
   id: PropTypes.string,
-  overEncumbrance: PropTypes.number,
-  overExpended: PropTypes.number,
 };
 
 BudgetInformation.defaultProps = {
@@ -188,8 +110,6 @@ BudgetInformation.defaultProps = {
   fiscalEnd: '',
   fiscalStart: '',
   name: '',
-  overEncumbrance: 0,
-  overExpended: 0,
 };
 
 export default BudgetInformation;