+++ src/components/Budget/BudgetView/BudgetView.js
@@ -13,8 +13,8 @@ import {
 import { ViewMetaData } from '@folio/stripes/smart-components';
 
 import ExpenseClasses from '../../../common/ExpenseClasses';
+import FinancialSummary from '../../../common/FinancialSummary';
 import { SECTIONS_BUDGET } from '../constants';
-import BudgetSummary from './BudgetSummary';
 import BudgetInformation from './BudgetInformation';
 
 const BudgetView = ({ budget, expenseClassesTotals, fiscalStart, fiscalEnd, fiscalYearCurrency }) => {
@@ -32,8 +32,8 @@ const BudgetView = ({ budget, expenseClassesTotals, fiscalStart, fiscalEnd, fisc
             id={SECTIONS_BUDGET.SUMMARY}
           >
             {budget.metadata && <ViewMetaData metadata={budget.metadata} />}
-            <BudgetSummary
-              budget={budget}
+            <FinancialSummary
+              data={budget}
               fiscalYearCurrency={fiscalYearCurrency}
             />
           </Accordion>
@@ -44,17 +44,13 @@ const BudgetView = ({ budget, expenseClassesTotals, fiscalStart, fiscalEnd, fisc
             <BudgetInformation
               allowableEncumbrance={budget.allowableEncumbrance}
               allowableExpenditure={budget.allowableExpenditure}
-              awaitingPayment={budget.awaitingPayment}
               budgetStatus={budget.budgetStatus}
-              encumbered={budget.encumbered}
               expenditures={budget.expenditures}
               fiscalEnd={fiscalEnd}
               fiscalStart={fiscalStart}
               fiscalYearCurrency={fiscalYearCurrency}
               name={budget.name}
               id={budget.id}
-              overEncumbrance={budget.overEncumbrance}
-              overExpended={budget.overExpended}
             />
           </Accordion>