+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## (IN PROGRESS)
 
+* Update Budget detail view to include more summary information. UIF-261.
 * Start Rollover process with some stub data. Refs UIF-171
 * "Based on" fields not required for encumbrance rollover when not in use. Refs UIF-274
 * Add personal data disclosure form. Refs UIF-272.